refactor: moved compiler and logger file directories
Basically traded spots to where they make most sense. While a compiler is a pretty essential part of Robo.js, it's only ever used in the CLI, which itself is also important. The logger, on the other hand, is used pretty much everywhere and thus is more appropriate for core.