patch: register process event listeners only when starting the robo
This makes sure that anything that may possibly reference this process.ts file in the /cli module doesn't accidentally register them!

This may happen because some CLI files reference the logger, which references the debug for the DEBUG_MODE constant, which references robo.ts.

Yes, the bleed can be handled differently, but this is important to handle at least this way now so it doesn't happen again in the future.