feat: support for subcommands and subcommand groups
It's now possible to create subcommands as well as subcommand groups!

The file structure works the exact way one would expect. Instead of making a command file, one can make a folder instead to set up subcommands in. The same is true for subcommand groups but with one more folder in.

Due to Discord limitations, it is not possible to register a command action for subcommands. For example, if you have a subcommand called "test cat", you cannot have a "test" command. The same is true for subcommand groups.