feat: new "portal" api
This essentially encapsulates command, context, and event records into one single container.

Think of this as a portal into the Robo's inner workings. As Gateway, if you will.

BREAKING CHANGE:
This replaces the `commands` export with `portal` instead. Update your code to use `portal.commands` if you previously relied on this.