feat(cli): experimental new file format for roboplay deployments
This would allow us to get rid of the `tar` dependency and structure uploads exactly how we want them! For example, improved process.env reads for times when the `.env` isn't there such as when using Doppler!

This is still experimental for now. Still working on getting decompression working just as well. It's nearly complete!