refactor(cli): handle parallel limits internally
The `p-limit` dependency relied on a special queue package which is overkill for our needs. It also didn't provide types.

The new implementation is technically computationally more expensive, but barely so at the current limit of 100. Even 1000 would be too little for it to make much of a difference.

This new implementation also supports TypeScript.