refactor(spirit): merged "command" and "event" fields; renamed "response" to "payload"
This makes things easier to manage as there was some overlap in functionality.

There's also a new `exec` function which words similar to `send`, except it waits until proper payload has been returned. It works well because we enforce spirits to always return a response, but we now also check payload for stop and restart payloads because that makes things work best without commands.

The state is now also saved, as a result of this.