feat(sage): error message channel redirects & custom messages
You can now specify an error channel to redirect sage errors to when not in production mode. This is useful when testing publicly and giving others access to debug info for a live bot. Running via `robo dev` is not required for this.

Additionally, it's now also possible to specify custom error messages to appear when something goes wrong during command interactions. This does not apply to events nor other potential async errors.