fix(spirits): handle missing id when running exec()
If the main Robo process fails for some reason, like if the global scope threw an error, a spirit id is never returned. This is by design.

However, the exec() function was never updated to account for that until now. This allows for graceful watcher rebuilds rather than an interrupted process!