feat: @roboplay/plugin-api
This plugin is a work in progress and pending Robo.js 0.9.0 release prior to publishing. Essentially, I'm working on decoupling the API functionality of the framework into a separate plugin.

The actual server logic was easy to separate. Now it's just the manifest build and portal loading that needs to be decoupled, which will require a major refactor.