feat: @roboplay/plugin-ai
Another work in progress plugin.

The purpose of this AI plugin is to enable seamless GPT integration, though other models may be plugged in on future updates.

A major feature here is that existing Discord.js commands can be acted upon by the AI itself. It does this via a hacky workaround involving creating mock objects. I'd like to use a safer alternative to this in the future.

All commands are available by default. Future update will include "Modes" which allow only a subset to be active at once.