refactor: removed built-in api server functionality in favor of new plugin
I was hesitant to add such a web server at first because of how little it had to do with most bots.

Instead, this functionality is slowly moving into a plugin. The stuff being removed in this commit already exists in the plugin. I'd like to move the build process into the plugin as well.