feat: new "robo add" command for installing plugins
It's now possible to install *and register* new plugins with one command!

If the plugin is already installed, it registers it. There's also a way to force these actions for whatever reason in case these smart checks are incorrect.

TODO: We still need to validate that the packages being installed are Robo.js plugins specifically. This way we avoid making a habit out of using this command for general packages and thereby cluttering the config directory.