feat(sage-cli): prompt to use latest version directly
Rather than tell users how to use the latest version, we now prompt with yes or no, and do it automatically.

This strategy works for this CLI and not the main Robo CLI because these are not repetitive actions. It'd be annoying to be prompted each time on the Robo. Plus, it's more important for Sage to be up-to-date.