patch(robo): auto delete outdated slash commands
Robo previously added and updated commands, but left it up to the devs to delete commands that no longer exist. This commit makes it so that we handle the entire process!

The `--force` flag will also do a full cleanup, even for commands that Robo.js was never aware of.