patch(robo): unified global instance for flashcore
This has been a common point of errors, especially for us working directly from this monorepo.

Recently we've found that plugins may have their own instance of Flashcore which is never initialized, thus leading to crashes. By moving the necessary initialized state to the `globalThis` object, we're able to fix this once and for all.