refactor(robo/cli): exclude `disableBot` from experimental warnings
Reason being, our support for activities is just as stable as bot support now and we're even reaching a point where the framework is used mostly for activities rather than bots.

Because of that, the "experimental" label is technically misleading, but we cannot yet remove the flag itself until the great refactor, so this is the next best alternate.