feat(plugin): @robojs/analytics (#272)
* WIP

* WIP

WIP

* plugin(analytics): Introducing analytics

* plugin(analytics): Refactored code

I refactored some part of the code and I am thinking about re-working it so it looks nicer too.

* plugin(analytics): amount fixed and new field

I added the referrer field so it can be used I also added check to make sure the amount sent to Plausible was always in a decimal format

* plugin(analytics): deleted robo build.

* chore(engines): reworked some code

I just reworked some code so it looks a bit cleaner.

* feat(google): added guards for params

I added guards so the param object is not filled with undefined

* fix(analytics): requested changes

Concerning name, yes it is necessary, the thing is page parameter and name are different where page is a segment of a link or full blown URL
name is the name of the event to register: subscribed etc

* chore(analytics): shortened check

* chore(analytics): smol monorepo adjustments

Just to get it working with our normal setup properly. Ya know, synergy and such. owo

* chore(analytics): added author and contributor credits

* chore(analytics): updated setup warning verbiage

* feat(templates): discord bot template using the analytics plugin

* docs(analytics): updated analytics readme

* refactor(analytics): moved analytics plugin directory

* chore(analytics): env correction & usage

* refactor(analytics): renamed some files & moved types

* chore(analytics): cleaned up google analytics engine

* feat(templates): example analytics usage in template

* refactor(analytics): refined plausible implementation

* chore(analytics): updated docs

* feat(analytics): seed

---------

Co-authored-by: Pkmmte Xeleon <hi@pkmmte.com>