Refactor WebSocket Server to Enhance Connection Handling and State Management (#326)
* refactor: optimize and modularize WebSocket server logic

* chore(sync): minor type & syntax adjustments

The `key` param in `broadcastUpdate` should be an optional `string[]` rather than `string`. The `cleanKey` is already the pure string version and `key` is meant to be an array for a useEffect-like dependencies array syntax.

Other than that, I just made two minor adjustments for our syntax preference. (heartbleed ptsd & optional chaining)

---------

Co-authored-by: Pkmmte Xeleon <hi@pkmmte.com>