chore(robo): removed pairing code requirement
Apparently it was causing too much friction for users and not everyone knew what to do.

Keep in mind that this was an already overkill security measure intended to guard against real-time attacks. You still have to have the generated token, which expires after use or a period of time, and obviously also need the login credentials.

The pairing code was an extra measure on top of that to prevent someone sniffing your Internet packets from interjecting themselves into the process and making you log in to their account instead. However, we already show the account that was logged into the CLI already, so there's security measures to prevent this as well that will persist.