feat(robo): ability to define user installs on a per-command basis w/ defaults
'Tis now possible to pass a `integrationTypes` array in slash command and context command config objects to define whether or not that should be user installable. No need to make everything user-installable anymore. This also opens up the ability to make commands that are _only_ user-installable.

It also includes a new `defaults.integrationTypes` field in the global config if you'd rather define this once rather than per command.